♻️ refactor(acceptance): split the reject modal into a model and two layouts
The modal had grown to 803 lines holding a state container, a desktop
tree, a phone tree and the pure helpers behind both. Worse, the desktop
half still carried six `md` branches and three `@media (width <= 767px)`
blocks that no phone can reach: the file returns the phone layout early,
so everything after it runs with `md === true`. Dead responsive code in
a file that is half responsive code is how the zoom bug hid.

Now:

- `useRejectReview` owns the whole model — draft, regions, attachments,
  zoom, the submit rule — so the two layouts can never disagree about
  what a submittable reject is.
- `DesktopEvidenceReview` and `MobileEvidenceReview` are siblings taking
  that model. Neither mentions a breakpoint.
- `EvidenceStage` is the shared zoom/pan surface. It owns its scroll node
  because that node IS the pan surface and the width zoom multiplies, and
  measuring a node you do not render is what left an observer watching a
  detached element.
- `RegionNotes` holds both note lists over one row atom; `rejectDraft`
  and `rejectModalShell` take the pure storage and chrome helpers.

`CheckRejectModal` is 66 lines: read the breakpoint, build the model,
pick a layout. No behaviour change.

Verified at 390x844 and 1440x900 against a local full-stack dev server:
the phone still circles, resizes, zooms 352 to 706px and lists notes
under the stage; the desktop keeps its thumbnails, floating zoom pill,
320px note panel and footer.
